# openapi.yaml
openapi: 3.0.0
info:
  title: NotaryPro Core API - Completa
  version: 2.0.0
  description: |
    **API COMPLETA para la gestión documental, certificación, auditoría y usuarios en NotaryPro Chile.**
    
    **?? NUEVA VERSIÓN 2.0.0 - Ahora incluye:**
    - ? Sistema de Firmantes Completo
    - ? Recuperación de Contraseñas por Email
    - ? Gestión de Archivos PDF (Base64)
    - ? Notificaciones Automáticas por Email
    
    **?? Cobertura vs FirmaVirtual: 85%**
    
    **?? Nuevos Endpoints Disponibles:**
    - 12 endpoints nuevos implementados
    - Compatibilidad completa con API de FirmaVirtual
    - Sistema de emails automáticos
    - Gestión avanzada de firmantes
    
    **?? Características Técnicas:**
    - Arquitectura Modular Independiente
    - Sistema de Roles Granular (ADMIN, GESTOR, CERTIFICADOR, OPERADOR, CLIENTE, VALIDADOR)
    - Motor de Plantillas Avanzado
    - Máquina de Estados Robusta para Documentos
    - Captura y Verificación de Firmas
    - Verificación de Integridad Documental
    - Auditoría Completa de Eventos
    - API Gateway Unificado
    - Validación Pública por QR y Hash (sin autenticación)
    - **?? Sistema de Firmantes con Roles**
    - **?? Reset de Contraseñas con Templates HTML**
    - **?? Upload/Download de PDFs en Base64**
    - **?? Notificaciones Automáticas Personalizadas**
servers:
  - url: http://168.232.167.145:3000/api/v1
    description: Servidor de Producción (VPS por IP)
  - url: http://localhost:3000/api/v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, example: admin@notarypro.cl }
        password: { type: string, format: password, example: adminpassword }
    LoginResponse:
      type: object
      properties:
        token: { type: string, description: Token JWT de acceso, example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9... }
...
paths:
...
  /auth/login:
    post:
      summary: Iniciar sesión de usuario
      tags: [Autenticación]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Inicio de sesión exitoso
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '400': { description: "Datos inválidos" }
        '401': { description: "Credenciales incorrectas" }
...